refactor(web): extract FileViewer's MarkdownViewer into the file-viewer slice
Cluster Q: mode/autosave/image-paste hook, shiki-highlight + code-block-copy
hook, anchor-based scroll-sync hook, wired + dumb components. Folds the
sibling markdown-scroll-sync.ts module into rules.ts (pure) and a new
MarkdownEditorMeasurePort (the one bare-document function). Two new provider
bridges: markdown-code-blocks.ts and theme-watch.ts. Reuses the existing
ShareLinkClipboardPort instead of adding a redundant clipboard adapter.
FileViewer.tsx 7235 -> 6431 lines.
